3.88->4.00, compilation tested up to 5.12.0-rc5 kernel
Note: this number DO NOT correspond to mats 400,
and mats 400 is saying that driver version is too new.

To make it work with mats400 an ugly hack was added to
esc_mods_get_api_version - if executable is named mats or mats 400 -
a fake older driver version is reported (and older mats is working fine).

Also set_fs(KERNEL_DS) is commented out for working on modern kernels,
maybe it will break some scenarios.